# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Robot Fleet API
  description: Public API for QA engineers to practice API testing, automation,
    contract testing, performance testing, security testing and modern backend
    quality engineering.
  version: 1.0.0
  contact:
    name: Robot Fleet API
    url: https://github.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register a new user account
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  minLength: 2
                  maxLength: 120
                  type: string
                email:
                  format: email
                  type: string
                password:
                  minLength: 8
                  maxLength: 72
                  type: string
              required:
                - name
                - email
                - password
        required: true
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  name:
                    type: string
                  email:
                    format: email
                    type: string
                  role:
                    anyOf:
                      - type: string
                        enum:
                          - ADMIN
                      - type: string
                        enum:
                          - MANAGER
                      - type: string
                        enum:
                          - OPERATOR
                      - type: string
                        enum:
                          - VIEWER
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - ACTIVE
                      - type: string
                        enum:
                          - INACTIVE
                      - type: string
                        enum:
                          - SUSPENDED
                  createdAt:
                    format: date-time
                    type: string
                  updatedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - name
                  - email
                  - role
                  - status
                  - createdAt
                  - updatedAt
  /api/v1/auth/login:
    post:
      summary: Authenticate and obtain access/refresh tokens
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  format: email
                  type: string
                password:
                  minLength: 1
                  type: string
              required:
                - email
                - password
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  expiresIn:
                    type: integer
                required:
                  - accessToken
                  - refreshToken
                  - expiresIn
  /api/v1/auth/refresh:
    post:
      summary: Exchange a refresh token for a new token pair
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required:
                - refreshToken
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  expiresIn:
                    type: integer
                required:
                  - accessToken
                  - refreshToken
                  - expiresIn
  /api/v1/auth/logout:
    post:
      summary: Revoke a refresh token
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required:
                - refreshToken
        required: true
      responses:
        "204":
          description: Default Response
  /api/v1/auth/me:
    get:
      summary: Get the currently authenticated user
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  name:
                    type: string
                  email:
                    format: email
                    type: string
                  role:
                    anyOf:
                      - type: string
                        enum:
                          - ADMIN
                      - type: string
                        enum:
                          - MANAGER
                      - type: string
                        enum:
                          - OPERATOR
                      - type: string
                        enum:
                          - VIEWER
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - ACTIVE
                      - type: string
                        enum:
                          - INACTIVE
                      - type: string
                        enum:
                          - SUSPENDED
                  createdAt:
                    format: date-time
                    type: string
                  updatedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - name
                  - email
                  - role
                  - status
                  - createdAt
                  - updatedAt
...
servers:
  - url: /
    description: Current server
...